Replay Quickalgo calls when the context visualization is unloaded and reloaded (e.g. when we change from desktop to mobile
