+++ frontend/player/sagas.ts
@@ -17,10 +17,10 @@ import {Bundle} from "../linker";
 import {makeContext, QuickalgoLibraryCall, StepperContext} from "../stepper/api";
 import {App, Codecast} from "../index";
 import {ReplayApi} from "./replay";
-import {quickAlgoLibraries} from "../task/libs/quickalgo_librairies";
+import {quickAlgoLibraries} from "../task/libs/quickalgo_libraries";
 import {ActionTypes as AppActionTypes} from "../actionTypes";
 import {taskLoad} from "../task";
-import {inputBufferLibTest, outputBufferLibTest, PrinterLibActionTypes} from "../task/libs/printer/printer_lib";
+import {inputBufferLibTest, PrinterLibActionTypes} from "../task/libs/printer/printer_lib";
 import {RECORDING_FORMAT_VERSION} from "../version";
 import {getCurrentImmerState} from "../task/utils";
 import {createDraft, finishDraft} from "immer";
