+++ frontend/stepper/js/BlocklyEditor.tsx
@@ -1,6 +1,6 @@
 import React, {useEffect, useRef} from "react";
 import {useAppSelector} from "../../hooks";
-import {quickAlgoLibraries} from "../../task/libs/quickalgo_librairies";
+import {quickAlgoLibraries} from "../../task/libs/quickalgo_libraries";
 import {ObjectDocument} from "../../buffers/document";
 import {BlockDocumentModel} from "../../buffers";
 import {ActionTypes} from "../../buffers/actionTypes";
